cdk/requirements-dev.txt
boto3>=1.36.14
PyJWT
django
pytz
requests
transformers
aws-cdk.aws-servicecatalogappregistry-alpha>=2.178.1a0
aws-solutions-constructs.aws-cloudfront-s3>=2.76.0
aws-cdk.aws-scheduler-alpha>=2.178.1a0
aws-cdk.aws-scheduler-targets-alpha>=2.178.1a0
aws-cdk.aws-lambda-python-alpha>=2.178.1a0
aws-cdk.aws-cognito-identitypool-alpha>=2.178.1a0
aws-lambda-powertools[all]